tests for passing around of both ipqs and mergers complete
it should now be impossible to pass the wrong merger or ipq without immediate test failure or mallicious input.

bug fixing tools for finding this sort of bug have been added: printers and raise statements. they are Very hard to track down, but should be substantially easier now, using the built in tools and warnings