bug found - if all m values are deleted, inf recursion
a bug has been found but not fixed - a relatively common case when k=2 - if all m values are in a known run, the code currently handles all known m values (aka nothing), and then infinitely recurses on the same input.

about to implement a solution: test for absence of m values - in that case, there must be a run. merge on the bigger run, recurse on the rest.